farm.lua: Add conditional right and left turns
This commit adds conditional right and left turns in the `goRight` and `goLeft` functions respectively. Previously, these functions only executed the `turtle.turnRight()` or `turtle.turnLeft()` once, without considering a specific condition. Now, if the input `x` is equal to 2, the turtle will make an additional turn in the corresponding direction.